LivingRoom
Added the couch. Changed the way that the archway's bottom vertices' were lining up with the floor and tiles by moving the bottom of the floor to be uniform instead of slanted and moved the bottom vertices of the archway down to full in the gaps. 

I need to bevel and make the complete the challenge for Lab 3 before moving onto Lab 4 and getting that done.