Living Room Challenge 3 - Picture Frame Next
Couch and Rug are done, where I last saved while working on it tonight before my computer just restarted and removed the TV set that I had. I did the rug first so I could understand how to make the picture next. 
To work on next: 
Create the Picture for challenge 3 and get that done 
Create the Lamp for Lab 4
Do the challenge for Lab 4 - the TV