Progess Update
Got most of the modeling done - i think the only thing I can think of to do still is the flowers and the fruit - I am speeding through the video on yt to see how he did those.

Started UV mapping a good chunk of stuff - when I select all am quickly able to see what still needs to be done and what I am working on still.

Next up is to do the modeling for the fruit and leaves - get some basic lighting in as well.

Then it will be onto UV mapping everything else then getting it all into substance painter