Modeling 100% and color
Got color added for everything - even if it was just a white lambert to be a place holder until I get the substance painter images in. Got the pictures in for all of the different menus/spicerack. Added a very basic global light to render the scene.

Added/Changed the modeling image and render image.

Updated what I need to do next as well